pyproject.toml
[project]
name = "lg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "faiss-cpu>=1.12.0",
    "sentence-transformers>=5.1.2",
    "langchain-huggingface>=1.0.1",
]
